+++ src/types/queries.ts
@@ -15,11 +15,24 @@ export interface IQueryRequest<D = undefined, T extends TQueryType = TQueryType>
 
 export type TQueryRequest<D = undefined, T extends TQueryType = TQueryType> = IQueryRequest<D, T>;
 
-export type TQuerySubscriptionMetadata = Array<[string | symbol, TQueryType]>;
+export type TQuerySubscriptionMetadata = Array<[ string | symbol, TQueryType ]>;
 
 export interface IQueryResponse<D = undefined, T extends TQueryType = TQueryType> {
+  /**
+   * Query key type for search matching.
+   */
   type: T;
+  /**
+   * Query requested data params.
+   */
   data: D;
+  /**
+   * Query emit timestamp.
+   */
+  timestamp: number;
+  /**
+   * Query answerer.
+   */
   answerer: TDreamstateWorker;
 }