+++ tests/assets/queries.ts
@@ -1,4 +1,4 @@
-import { ContextWorker, ContextManager, OnQuery, QueryRequest, QueryResponse } from "@Lib";
+import { ContextWorker, OnQuery, QueryRequest, QueryResponse } from "@Lib";
 
 export enum EQuery {
   SYNC_BOOLEAN_QUERY = "SYNC_BOOLEAN_QUERY",
@@ -6,7 +6,6 @@ export enum EQuery {
   ASYNC_STRING_QUERY = "ASYNC_STRING_QUERY",
   SYNC_EXCEPTION_QUERY = "SYNC_EXCEPTION_QUERY",
   ASYNC_EXCEPTION_QUERY = "ASYNC_EXCEPTION_QUERY",
-  ASYNC_INTERCEPTOR_QUERY = "ASYNC_INTERCEPTOR_QUERY",
   UNDEFINED_QUERY = "UNDEFINED_QUERY"
 }
 
@@ -15,11 +14,8 @@ export type TAsyncStringQuery = QueryRequest<string, EQuery.ASYNC_STRING_QUERY>;
 export type TSyncBooleanQuery = QueryRequest<void, EQuery.SYNC_BOOLEAN_QUERY>;
 export type TAsyncExceptionQuery = QueryRequest<void, EQuery.ASYNC_EXCEPTION_QUERY>;
 export type TSyncExceptionQuery = QueryRequest<void, EQuery.SYNC_EXCEPTION_QUERY>;
-export type TAsyncInterceptorQuery = QueryRequest<void, EQuery.ASYNC_INTERCEPTOR_QUERY>;
 
-export class RespondingContextManager extends ContextManager<object> {
-
-  public context: object = {};
+export class RespondingWorker extends ContextWorker {
 
   @OnQuery(EQuery.ASYNC_EXCEPTION_QUERY)
   public async onAsyncExceptionQuery(queryRequest: TAsyncExceptionQuery): Promise<never> {
@@ -48,18 +44,16 @@ export class RespondingContextManager extends ContextManager<object> {
 
 }
 
-export class RespondingWorker extends ContextWorker {
+export class RespondingDuplicateWorker extends ContextWorker {
 
-  @OnQuery(EQuery.ASYNC_INTERCEPTOR_QUERY)
-  public async onAsyncInterceptorQuery(queryRequest: TAsyncInterceptorQuery): Promise<number> {
-    return Math.random();
+  @OnQuery(EQuery.ASYNC_NUMBER_QUERY)
+  public async onAsyncNumberQuery(queryRequest: TAsyncNumberQuery): Promise<number> {
+    return -1;
   }
 
 }
 
-export class RequestingContextManager extends ContextManager<object> {
-
-  public context: object = {};
+export class RequestingWorker extends ContextWorker {
 
   public async sendUndefinedQuery(): Promise<QueryResponse<any>> {
     return this.sendQuery({ type: EQuery.UNDEFINED_QUERY });
@@ -83,10 +77,6 @@ export class RequestingContextManager extends ContextManager<object> {
 
   public async sendAsyncStringQuery(data: string): Promise<QueryResponse<string>> {
     return this.sendQuery({ type: EQuery.ASYNC_STRING_QUERY, data });
-  }
-
-  public async sendInterceptorQuery(): Promise<QueryResponse<number>> {
-    return this.sendQuery({ type: EQuery.ASYNC_INTERCEPTOR_QUERY });
   }
 
 }